+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Fixed
+- Allow format 12 subtables with *Unicode full repertoire* in `cmap`.
 
 ## [0.2.1] - 2019-08-12
 ### Fixed